refactor: Shims are done
- Separated some of the shims into their own files
