+++ typescript/vrotsc/src/compiler/transformer/declaration.ts
