+++ typescript/vrotsc/src/compiler/transformer/helpers/README.md
@@ -11,3 +11,9 @@ It is not specific to Aria Orchestrator.
 Contains a set of functions to work with TypeScript source files.
 The source helper can modify the AST.
 It is not specific to Aria Orchestrator.
+
+## VROES.ts
+
+Contains helpers specific for VROES.
+VROES is a utility tool for imports/exports.
+This is not the tool itself, but helpers for it
