+++ typescript/vrotsc/src/compiler/transformer/helpers/VROES.ts
@@ -1,0 +1,7 @@
+export const SCRIPT_VRO_GLOBAL = "__global";
+export const SCRIPT_VROES_VAR = "VROES";
+export const SCRIPT_VROES_MODULE = "VROES";
+export const SCRIPT_VROES_CACHE = "__VROES";
+export const SCRIPT_HELPER_MODULE = "tslib";
+export const SCRIPT_LAZY_IMPORT_NAME = "importLazy";
+export const SCRIPT_VRO_MODULE_PACKAGE = "com.vmware.pscoe.library.ecmascript";
