+++ typescript/vrotsc/src/compiler/transformer/helpers/source.ts
@@ -49,3 +49,4 @@ export function printSourceFile(sourceFile: ts.SourceFile): string {
 	});
 	return printer.printFile(sourceFile);
 }
+
